検査実施件数: Rename components/TimeStackedBarChart.vue to components/index/CardsReference/TestedNumber/Chart.vue
